MUSARK-940 Adding events table
